Corrections after Code Review
